Added block spawn probability
- Blocks inside structures can have attribute for placement probability
- Code clean-up
- Adjusted per-face brightness in frag shader
- New noise for block placement probability

+++ src/world/structureDB.hpp
@@ -5,7 +5,7 @@
 #include <string>
 #include <unordered_map>
 
-using StructureLayer = std::vector<std::vector<uint8_t>>;
+using StructureLayer = std::vector<std::vector<uint16_t>>;
 
 class Structure {
 public:
